Add files via upload
Course project for Database systems.

The project was divided into two sides: student and staff. This is the staff side.

Staff is then further divided into staff types: FYP, Finance, OneStop and Director.

The Director has a singular log-in and can add new staff by specifying a type.
OneStop can view all requests and their status and notifies the student if the request has been approved and processed.
FYP department can approve/reject a student's request.
Finance can also approve/reject a student's request.
Both Finance and FYP permissions are required to be in an accepted state to move forward with the application.